Introducing data_source table
Idea is to have an entry for each legacy database,
latest NCBI Entrez import, etc. This can take over
the planned source entry date added field.

Note, having to introduce auto-incremented primary
IDs for this - so might next switch to that in place
of the MD5 for linking the sequence table.