Species name validation at import - WIP
Can't deal with synonyms where we only have the new name in our database.
See a few of these in the NCBI import, e.g.

- Pythium undulatum, aka Phytophthora undulata
- Peronophythora litchii, aka Phytophthora litchii

Needs refactoring to handle those NCBI FASTA entries where the species is
not one word, e.g.

- Phytophthora aff. asparagi
- Phytophthora cf. rosacearum PB-2013
- Phytophthora sp. personii
- Phytophthora taxon Pgchlamydo

Right now this only has logic for dropping words (since with the
legacy FASTA files there are sometimes extra bits), not adding words.

For the legacy import the clades are a problem - all the pre-loaded
NCBI taxonomy entries do not have a clade, so most of the entries in
the legacy file do not match if we include the clade.