Use predictable filenames within temp dir; add --cpu
This follows the approach I used in the prepare-reads
command, and opens up making debugging easier by using
a pre-set temp folder or optionally not removing the tmp
after a failure when debugging.

Also defines --cpu for the classifier (not using it for
the identity method calling hmmscan as it did not seem
to speed things up).