Hamming Loss in thapbi assess
TO DO: How do we decide the number of species (number of classes),
which is a divisor in the Hamming Loss definition? Currently using
the species count as for our confusion matrix and report - which
is the union of expected or predicted species (and thus depends
indirectly on the classifier output).

Could argue for using number of species in the DB, given that is
the maximum number of classes which could be predicted.

Currently getting (for our single isolate controls), Hamming Loss
values of:

0.9667 - Say all samples have all control isolates (mostly FP)
0.1042 - Just say P. austrocedrae;infestans;pseudosyringae
0.0552 - Just say P. infestans (some TP, loads of FP/FN)
0.0333 - Just say no predictor (zero TP, FP, but maximum FN rate)

We get close to a 100% loss with the trivial predictor which says
everything is present, meaning it scores mostly FP with a few TP.

The next worse score at about 10% loss is a trivial predictor which
just says it is the three most common single isolate species on the
control plate - this get penalize for typically making three FP on
every example.

The identity, swarm and blast are better (lower), exactly values
depend on the DB of course. These values are all very low - which
is why I am reporting to 4dp.