1bp difference classifier
Based on the 'identity' classifier, if a prepared read matches
a database sequence, uses the taxonomy hit(s) from that just like
the identity classifier.

However, if that fails, uses a dictionary to see if the prepared
read is only 1bp away from a DB entry. If so, it reports the
taxonomy results for those 1bp different DB entries.

Considers 1bp substitutions, deletions, and insertions (including
adding 1bp to the start, or the end).