Re-order columns, fixed then variable columns
May not be as human readable, but putting the variable
per-sample columns last ensures all the other columns
are in fixed locations - useful for downstream processing.

Also adjusted order (MD5, predictions, seq) to suit Excel.
As the predictions are often empty, in Excel the previous
column will be shown spilling into it (if it is long).