Record DB species list in classifier output; use to report TN
Previously we were failing to report all the true negative (TN)
predictions, because at the assess step we didn't know about
all the unreported species which were correctly not called.

In order to do this, we need to report these negative predictions
in the classifier output somehow. One option would be to report
a sparse matrix, one row for sequence (as now), with one column
per species - that would be a little unwieldy with lots of empty
columns.

Instead, I have recorded the DB species line embedded in the
TSV header line.

Note that while I could have recorded the DB species count, that
would not have let us spot cases where the true result as not in
the DB (so the classifier couldn't report it) which ought to get
reported as false negatives (and/or false positives).