Apply fixed trimming before calling hmmscan
This assumes that hmmscan will still find the HMM on the shorter
trimmed sequence (which is will unless the fixed trimming has
taken off too much of the sequence, in which case we may have
had trouble classifying the truncated sequence anyway).

The motivation for this is to then apply the minimum abundance
threshold before hmmscan, and thus we only have to give a small
fraction of the trimmed reads to it - making this much faster.

Initial tests on a single sample suggest ~2mins instead of ~10min,
which suggests a 96-sample plate would take ~4hrs instead of ~16hrs.