Apply color banding to sample columns in Excel
Looks at the displayed metadata to try to find a field which is
grouped into self-contained blocks (after the column sorting)
and thus makes sense to use for color banding where we reused
colors (not a color key).

If none of the metadata fields work as is, tries the first 'word'
for each (breaking on spaces, minus signs or underscores) which
is designed with our sample naming scheme in mind.