Refactor filter_for_ITS1 to return all matches, not span
I think it is actually a return to an older version of the function
behaviour?

With this change, the intent is the DB import tools will take all
the matches in a single sequence (e.g. tandem repeats in genome).
Currently however, just takes the first with a warning.

For the prepare-reads step, could reasonably abort, but will for
now accept the merged read as is.

The classify step will abort (don't expect to get two or more ITS1
regions via our Illumina approach due to read lengths).