Use a non-alphabetical cross reference list in metadata
The order of the three DNA mix entries can then be checked in
a test (alphabetical vs the order in the metadata column).

Note could already do that with the Site_N_Sample_X entries where
X ran from 1 to 10, and have been entered in logical order.

This might make a useful option to add, along side if we should
respect the order of the rows in the metadata table or not.