Reformat legacy curated trimmed ITS1 set v0.004 (27 Feb 2018)
$ ./control_a.py < Phytophthora_ITS_database_v0.004.fasta > ../Phytophthora_ITS1_curated.fasta

Original taken from Phytophora_ITS_database_v0.004.fasta (with
spelling corrected to use Phytophthora) from here:

https://github.com/peterthorpe5/public_scripts/blob/master/metapy/data/Phytophora_ITS_database_v0.004.fasta

Original commit was 27 Feb 2018, as part of a wider update to
the metapy pipeline:

https://github.com/peterthorpe5/public_scripts/commit/4e565ce013d9694015e93b355b3280cde3c08718

Contains 170 sequences with lengths from 136bp to 203bp.

This is a wholescale replacement of the older database.fasta,
with the most obvious difference being the sequences are now
trimmed to the ITS1 barcode region of interest.