Only use HMM for presence/absense, not trimming
Currently we use the HMM to trim on import to the DB,
and to trim in most of the classifiers to compare to
the DB. This seemed like a good idea for the perfect
identity or onebp classifiers, and does seem to work.

However, have noticed inconsistent HMM trimming as
the amount of flanking sequence changes - see GitHub
issue #192, seem worth revisiting issue #48.