v0.6.0 - Stop discarding normally conserved 32bp start of marker.
We had been ignoring this to date (treating it as part of the
left primer and removing it) since many of the published
Phytophthora ITS1 sequence omit this typically conserved
32bp region.

However, with recent work to drop the Phytophthora HMM filter
to consider more non-Phytophthora sequences, makes sense to
include the full sequenced region.

In order to maintain the existing curated set, where possible
they have been extended using the published upstream sequence.
Where this is not available, it has been lower case padded with
the expected TTTCCGTAGGTGAACCTGCGGAAGGATCATTA data - and the
accession version suffix changed to '.X' to indicate there was
a modification (X for extension).