v0.5.8: Correct start of AY659739.1 P. parsiana marker
CGCACCTAA... -> CCACACCTAA...

Confirmed with David Cooke: The currated set version of the
marker had one of the leading C missing. Furthermore, the
published sequence then had a G which on reviewing the original
traces ought to have been A, meaning this start now matches the
other two P. parsiana in our set.

Will look forward to AY659739.1 being updated as AY659739.2
on the NCBI collection later.