Tighten min/max length for NCBI import
Defaults are 100 and 1000, but now that we're including
the typically conserved 32bp region justified in raising
the minimum accordingly.

Moreover, need to avoid swelling the DB too much to cap
the memory (RAM) requirements for the onebp classifier.
The problem here are NCBI sequences where the primers
do not match (yet we are accepting them anyway untrimmed).