Simplified Redekar et al 2019 Suplementary Table 3
Based on the Excel file provided as part of:

    Redekar et al. (2019) Diversity of Phytophthora,
    Pythium, and Phytopythium species in recycled
    irrigation water in a container nursery.
    https://doi.org/10.1094/PBIOMES-10-18-0043-R

This has been reduced to just two columns, accession
(with dot version added where missing) and the given
species (which may not match the current annotation).
This and dropping the other columns gave duplicate
rows which have been removed. Additionally a missing
space after an 'sp.', and an inconsistent underscore
versus hypen was fixed in the species names.

This has been sorted by species, and then accession.

This is a plain text tab separated table, with 1454
rows (i.e. 1454 unique accessions, 430 unique species
names as listed).