Try FASTP without adapter trimming
As long as it can merge the reads, we don't care
about removing the adapters as we're going to do
PCR primer triming anyway.

This seems to halve the prepare-reads runtime!
(With no change on the dataset tried at -a 100)

However, FASTP documentation suggest removing them
helps with the pair merging.