Brute force Levenshtein edit-distance classifier
This takes the DB taxonomy for any perfect matches
(to species level if available), but failing that
takes the genus level taxonomy of the closest DB
entries within the maximum edit distance of 3.

Could tweak this with separate thresholds to build
the 1s2g classifier I've considered in the past.