Curated Nothophytophthora ITS1 sequences
Created with guidance from David Cooke.

As with the Phytophthora, most published sequences are
missing the start expected with our primers, and we have
had to assume it (added in lower case).

We have perfect matches to the extended N. intricata
sequence, and KP208466 agrees too - but more direct
confirmation for the other species would be preferred.