Update these expected results avoiding double counting
The old output was from running cutadapt separately
for each primer pair, and meant with overlapping
amplicons a read-pair for the long target would also
get reported for the short target as well.

The new output is from running cutadapt with all the
primer pairs at once, doing demultiplexing of the
amplicons. This way a read-pair can only be assigned
to a single amplicon target (the longer product in
the case of overlaps).

Thus the counts for Mini-16S, Mini-rbcL and trnL-P6-loop
all drop (no longer includes reads from the longer 16S,
rbcL and trnL-UAA products).