low-memory onebp and 1s3g etc implementation (#372)
Drop the onebp variant cloud which was problematic when the DB was very large (e.g. over 4000 sequences).

Instead checks the cloud of possible 1bp edits of each query sequence instead (which benefits from the pipeline changes in v0.9.9 to run the classifier on a non-redundant FASTA file), needing far less RAM.

The distance based classifiers (1s2g, ..., 1s5g) now explicitly call the onebp classifier initially, and thus benefit from these changes too.

Testing with our current default DB, memory usage drops from a sustained ~450MB to about ~55MB , and runtimes are about the same. Testing with a DB of over 100k entries is now possible (which was blocking adding a potential worked example).

This is expected to degrade performance if the classifier is called on multiple FASTA files with overlapping sequence content (as per the pipeline prior to v0.9.9).

Closes #346