python-Levenshtein --> rapidfuzz
The original python-Levenshtein library was not being
maintained with no pre-compiled wheels available on PyPI,
nor Python 3.10 support.

The fork https://github.com/maxbachmann/Levenshtein/ did
have wheels but changed the name on PyPI and is not (yet)
available on conda. However, it was now a wrapper for the
author's https://github.com/maxbachmann/RapidFuzz which
we can call directly.

RapidFuzz is availble on both PyPI (with wheels) and conda
(again, via conda-forge).