Fix k-mer threshold for finding spike-in sequences
The original was wrongly using 10 all the time, meaning
with 31-mers a match could be called with just a 40bp
match.

This was not picked up on our own spike-in datasets,
but did raise a false match on Palmer et al. (2018)
for a Saccharomyces cerevisiae sequence in BioMockStds
SRR7109326, where the last 40bp matched their mock_2.
Note their synthetic spike-in sequences were based
on Saccharomyces cerevisiae, so this makes sense.