No absolute/fractional threshold on synthetic/negative controls
I initially did not apply the fractional abundance
threshold to controls (at all) to avoid changes to
the behaviour of existing examples using absolute
threshold via negative controls.

There is also a concern if (some of the) controls are
at a different coverage depth to (some of the) samples.
e.g. if the control was low sequencing depth, a high
default absolute threshold could discard unwanted
reads which might have raised the fractional threshold.

This compromise ignored the requested absolute and/or
fractional thresholds for the controls which might
affect the other threshold.

It does mean if a control is entered as BOTH a negative
(-n) and synthetic (-y) control, it will be processed
with no threshold at all.