Apply half absolute & fractional abundance to controls.
This compromise applies half the absolute and/or half
the fractional thresholds to the controls when they
might affect the other threshold.

Consider an "empty" negative control mislabelled as a
synthetic control, or a failed synthetic control - you
could have just a handful of reads (say 6) and get an
artifically high inferred fractional abundance, e.g.
16.7% (i.e. 1/6), when clearly this is at the noise level.