Always show threshold and threshold-pool in summary reports
While the threshold is redundant if all the samples had the same
value (e.g. high absolute threshold, low fractional threshold),
and the threshold pool is redundant if all the samples where in
the same sub-folder, it should make it easier to document and
explain how the thresholding works if always shown.