Insert missing T or TT from start of NCBI seqs
Also require 32bp overlap with the 32bp leader
check. See issue #497.

$ grep -v "^>" 2022-07-05_ITS1_Oomycota_w32.fasta \
  | cut -c 1-32 | sort | uniq -c | sort -n | tail
   9 TTTCCGTAGGTGAACCTGCGGAAGGATCATTT
   9 tTTCCGTAGGTGAACTGCGGAAGGATCATTAC
  16 tTTCCGTAGGTGAACCGCGGAAGGATCATTAC
  20 CTTCCGTAGGTGAACCTGCGGAAGGATCATTA
  36 ttTCCGTAGGTGAACTGCGGAAGGATCATTAC
  48 TTTCCGTAGGTGAACTGCGGAAGGATCATTAC
  69 GTTCCGTAGGTGAACCTGCGGAAGGATCATTA
 123 tTTCCGTAGGTGAACCTGCGGAAGGATCATTA
 336 ttTCCGTAGGTGAACCTGCGGAAGGATCATTA
2902 TTTCCGTAGGTGAACCTGCGGAAGGATCATTA

This still allows substitutions (e.g. the G start)
and deletions (e.g. missing a C halfway along).

Note while we show the inserted T or TT in lower case,
this does not alter the accession to indicate the edit.

Also note the insertion is only made at the start of the
accession sequence - so will not be made if there is any
upstream sequence like a partial primer. For example,
Phytophthora MG553204.1, MG696526.1 and OM984702.1 used
to have the same sequence with a TT start. The first two
get extended to TTT, but OM984702.1 was not as published
sequence starts CAAGGTTCCGTAGGTGAACCTGCGGAAGGATCATTA.