Rebuild DB, update tests
Net increase in almost 100 Phytophthora, however...

Note for example these 3 Phytophthora used to have the same
sequence with a TT start: MG553204.1, MG696526.1 and OM984702.1
The first 2 are now extended to TTT, but OM984702.1 was not as
published sequence starts CAAGGTTCCGTAGGTGAACCTGCGGAAGGATCATTA
i.e. DB now has this amplicon twice, with TTT and TT at the start.