Make a single call to cdist for distance classifiers
Replace the variant cloud dictionary with a single
call to the RapidFuzz cdist function to compute all
the pairwise edit distances between the query and
database sequences in one go.

Also replaces the database sequence set with a dict
using the associated genus as value (saves a lot of
DB lookups for a modest memory usage).

Testing with 2k queries and 120k entries in DB,
sees to more than halve the classify run time for
the onebp and 1s3g classifiers.