Always apply -a and -f to controls in sample-tally
Regression from v0.13.2 to v0.13.3 allowed marginal
entries from controls to be accepted into the
'all reads' FASTA if below the given threshold but
above half of it (heuristic used in prepare-reads
to ensure hash -a setting didn't prevent autoincrease
to the -f setting, and vice versa).

Originally this fuzzy limit on controls was done in
prepare-reads, but overridden with the -a value in
the fasta-nr step.