Apply minimum total abundance in sample-tally
Regression from v0.13.2 to v0.13.3 allowed marginal
entries from controls to be accepted into the
'all reads' FASTA if below the given threshold but
above half of it (heuristic used in prepare-reads
to ensure that -a setting didn't prevent autoincrease
to the -f setting, and vice versa).

In switching from fasta-nr to sample-tally to make
the "all reads" FASTA, was missing the -t / --total
value in the pipeline (set to -a).