Dynamic limits on UNOISE dist and centroids
We can compute an upper bound on the distances (and
RapidFuzz can use this), and also exclude low abundance
centroids (so even less time computing distances).

Why? Near the end of the run, will have many low
abundance isolated nodes which will not be able
to adsorb the new node - can exclude them from
the distance matrix.

Makes a substantial speedup on soil_nematodes example,
with no change to the output (from 8mins to 2mins).