Modify the Phytophthora cinnamomi denoise example.
With a centroid abundance 5689583, the second sequence
at 90062 was almost 64x lower. It is now under 32x lower.

Previously this meant with d=2 (Levenstein), threshold
1:32, it was merged into the centroid by our UNOISE
implementation.

However, it appears USEARCH and VSEARCH see this as d=3
(substitution count) and therefore it was over the skew
threshold 1:128 and not merged.

By raising the abundance of the second sequence, even at
d=2 it will now not get merged and become a second centroid.

i.e. Our implementation and USEARCH/VSEARCH will agree,
giving two unique outputs.

This makes it clearer to see the effect of abundance vs
distance based greedy clustering.