Modify Phytophthora lacustris vs riparia denoise example
Very much the same change as just made to the
Phytophthora cinnamomi example.

With a centroid abundance 5689583, the second sequence
at 44733 was 127x lower. It has been increased to be
just under 32x lower.

Previously this meant with d=2 (Levenstein), threshold
1:32, they were merged into the centroid by our UNOISE
implementation.

However, it appears USEARCH and VSEARCH see this as d=3
(substitution count) and used the 1:128 threshold and
it was NOT merged.

By raising the abundance of the second sequence, even at
d=2 it will now NOT get merged and become a second centroid.

i.e. Our implementation and USEARCH/VSEARCH will agree,
giving two unique outputs.

This makes it clearer to see the effect of abundance vs
distance based greedy clustering.