Adding observed entry ON130151.X Peronospora gaponenkoae
This is a new sequence published 1 Jan 2024, and not yet in the
NCBI taxonomy.

This appears (with the typical 32bp leader) in multiple samples.

It ought to be added via Oomycota_ITS1_obs.fasta next time our
NCBI search results are refreshed, but there is something odd
going today with multiple Phytophthora nicotianae not being
returned by esearch/efetch.