Update NCBI search results.
This was a hassle. The current search seems to
miss a bunch of Phytophthora nicotianae ITS1
sequences (e.g. OP753715) but they are still
available and can be found via different search
terms. Specifically it seems they can't be found
via the organism search field - suggesting a
glitch in the Entrez index?

See also #587, would have needed to add ON130151
Peronospora gaponenkoae to the observed file but
did that yesterday via the curated file.