Stop using freeze-panes in Excel output
This was a good idea with the earlier simpler reports
and a large screen, but as we're tracking more sample
metadata like raw FASTQ counts, and often have user
supplied metadata, this can take up most of the screen.

Working on a laptop, I was finding myself hiding the
metadata columns in order to see the main counts.