Optional -k / --markers to pipeline and prepare-reads
Use case is sequencing runs using only a subset of the markers
defined in a multiple-marker database. e.g. Our historic MiSeq
plates using ITS1 when using a DB containing ITS1 and rps10
marker defintions.

Default remains looking for all the markers, and the warnings
for using a subset are currently not linked to the verbose mode.