Show rich progress bars for preparing reads and unoise-l
This seems to work reasonably well even with printing
debug/progress messages which then scroll up above the
progress bar which stays at the bottom of the screen.
(Might be because we only print whole lines?)

Will be harder to add this for long-running external
tools like VSEARCH without perhaps monitoring their
output as it is written?