Tweak how rich-argparse shows command line snippets & filename
Previously the -unoise and --cluster_unoise bits alone got highlighted
in the args style. Switching to back-ticks gets the snippet formatted
as a syntax block which is bold by default (if supported in the terminal).
However, the argument within that still got the cyan styling. That might
be a bug, but the simple workaround is to use the same style for backticks.

This also using back-ticks for filenames and sequencing naming snippets.